ai_microservice/requirements.txt
fastapi
uvicorn[standard]
requests
python-multipart
# torch
# transformers
# sentence-transformers
# sacremoses
# scipy
# openai-whisper 
# librosa
pillow
google-genai
